adjustments and fixes to derived table script
